sketches working
